Change name classes and directories and Given phrases of scenarios
