+++ build.gradle
@@ -49,7 +49,7 @@ task cucumberCli() {
                     '--plugin', 'html:build/cucumber/cucumber-report.html',
                     '--plugin', 'json:build/cucumber/cucumber.json',
                     '--plugin', 'rerun:build/cucumber/rerun.txt',
-                    '--glue', 'salesforce.scenarios',
+                    '--glue', 'rest.salesforce.com.scenarios',
                     'src/test/resources/features']
             ignoreExitValue = true
         }
@@ -60,7 +60,7 @@ task executeFeatures(type: Test) {
     useTestNG()
     options {
         useDefaultListeners = false
-        jvmArgs = ['-Dcucumber.glue=salesforce']
+        jvmArgs = ['-Dcucumber.glue=rest.salesforce.com.scenarios']
         jvmArgs = ['-Dcucumber.features=src/test/resources/features/']
         jvmArgs = ['-Dcucumber.plugin=pretty,' +
                            'html:build/cucumber/cucumber-report.html,' +
