+++ src/main/java/salesforce/auth/Authentication.java
@@ -8,7 +8,7 @@
 package salesforce.auth;
 
 import org.apache.http.HttpHeaders;
-import salesforce.config.CredentialsConfig;
+import salesforce.config.Credentials;
 import salesforce.config.HeaderValue;
 import salesforce.entities.Token;
 
@@ -20,16 +20,16 @@ public class Authentication {
     public static Token getAuth() {
         return
                 given().urlEncodingEnabled(true)
-                        .param(CredentialsConfig.USERNAME1.getEnumValue(), dotenv.get(CredentialsConfig.USERNAME1.getEnumName()))
-                        .param(CredentialsConfig.PASSWORD.getEnumValue(), dotenv.get(CredentialsConfig.PASSWORD.getEnumName()))
-                        .param(CredentialsConfig.CLIENT_ID.getEnumValue(), dotenv.get(CredentialsConfig.CLIENT_ID.getEnumName()))
-                        .param(CredentialsConfig.CLIENT_SECRET.getEnumValue(), dotenv.get(CredentialsConfig.CLIENT_SECRET.getEnumName()))
-                        .param(CredentialsConfig.GRANT_TYPE.getEnumValue(), CredentialsConfig.PASSWORD.getEnumValue())
+                        .param(Credentials.USERNAME1.getEnumValue(), dotenv.get(Credentials.USERNAME1.getEnumName()))
+                        .param(Credentials.PASSWORD.getEnumValue(), dotenv.get(Credentials.PASSWORD.getEnumName()))
+                        .param(Credentials.CLIENT_ID.getEnumValue(), dotenv.get(Credentials.CLIENT_ID.getEnumName()))
+                        .param(Credentials.CLIENT_SECRET.getEnumValue(), dotenv.get(Credentials.CLIENT_SECRET.getEnumName()))
+                        .param(Credentials.GRANT_TYPE.getEnumValue(), Credentials.PASSWORD.getEnumValue())
                         .header(HttpHeaders.ACCEPT, HeaderValue.APP_JSON.get())
                         .header(HttpHeaders.CONTENT_TYPE, HeaderValue.APP_X_FORM.get())
                         .log().all()
                         .when().
-                        post(dotenv.get(CredentialsConfig.TOKEN_URL.getEnumName()))
+                        post(dotenv.get(Credentials.TOKEN_URL.getEnumName()))
                         .as(Token.class);
     }
 }
