+++ src/main/java/salesforce/config/Credentials.java
@@ -7,7 +7,7 @@
  */
 package salesforce.config;
 
-public enum CredentialsConfig {
+public enum Credentials {
     USERNAME1("USERNAME1", "username"),
     PASSWORD("PASSWORD", "password"),
     CLIENT_ID("CLIENT_ID", "client_id"),
@@ -19,7 +19,7 @@ public enum CredentialsConfig {
     private String name;
     private String value;
 
-    CredentialsConfig(final String name, final String value) {
+    Credentials(final String name, final String value) {
         this.name = name;
         this.value = value;
     }
