+++ src/test/java/rest/salesforce/com/scenarios/runner/TestRunner.java
@@ -5,7 +5,7 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala.
  */
-package salesforce.scenarios.runner;
+package rest.salesforce.com.scenarios.runner;
 
 import io.cucumber.testng.AbstractTestNGCucumberTests;
 import org.testng.annotations.AfterSuite;
