+++ src/test/java/rest/salesforce/com/tests/CommonTest.java
@@ -5,7 +5,7 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package salesforce.tests;
+package rest.salesforce.com.tests;
 
 import salesforce.auth.Authentication;
 import salesforce.entities.CreatedResponse;
