+++ src/test/java/rest/salesforce/com/tests/IndividualTest.java
@@ -5,90 +5,93 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package salesforce.tests;
+package rest.salesforce.com.tests;
 
 import api.ApiMethod;
-import api.ApiRequestManager;
+import salesforce.config.Request;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import salesforce.config.Endpoints;
 import salesforce.entities.CreatedResponse;
-import salesforce.entities.Account;
+import salesforce.entities.Person;
 import org.apache.http.HttpStatus;
-import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
+import org.testng.annotations.AfterMethod;
 import java.util.HashMap;
 import java.util.Map;
 
-public class AccountTest extends CommonTest {
+public class IndividualTest extends CommonTest {
 
     @BeforeMethod(onlyForGroups = {"get", "update", "delete"})
     public void create() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
-        Account account = new Account();
-        account.setName("First Account");
+        Person person = new Person();
+        person.setFirstName("Pepito");
+        person.setLastName("Flores");
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNTS.get(), pathParams, account, ApiMethod.POST);
+        apiResponse = Request.execute(Endpoints.PEOPLE.get(), pathParams, person, ApiMethod.POST);
         createdResponse = apiResponse.getResponse().as(CreatedResponse.class);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_CREATED).log().body();
     }
 
     @Test(groups = "get")
-    public void getAllAccountTest() throws JsonProcessingException {
+    public void getAllIndividualTest() throws JsonProcessingException {
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNTS.get(), new HashMap<String, String>(), ApiMethod.GET);
+        apiResponse = Request.execute(Endpoints.PEOPLE.get(), new HashMap<String, String>(), ApiMethod.GET);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_OK).log().body();
     }
 
     @Test(groups = "post")
-    public void createAAccountTest() throws JsonProcessingException {
+    public void createAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
-        Account account = new Account();
-        account.setName("Created Account ");
+        Person person = new Person();
+        person.setFirstName("Pepito");
+        person.setLastName("Flores");
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNTS.get(), pathParams, account, ApiMethod.POST);
+        apiResponse = Request.execute(Endpoints.PEOPLE.get(), pathParams, person, ApiMethod.POST);
         createdResponse = apiResponse.getResponse().as(CreatedResponse.class);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_CREATED).log().body();
     }
 
     @Test(groups = "get")
-    public void getAAccountTest() throws JsonProcessingException {
+    public void getAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), createdResponse.getId());
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, ApiMethod.GET);
+        apiResponse = Request.execute(Endpoints.PERSON.get(), pathParams, ApiMethod.GET);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_OK).log().body();
     }
 
     @Test(groups = "update")
-    public void updateAAccountTest() throws JsonProcessingException {
+    public void updateAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), createdResponse.getId());
-        Account account = new Account();
-        account.setName("Updated Account");
+        Person person = new Person();
+        person.setFirstName("Pedrito");
+        person.setLastName("Fantasy");
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, account, ApiMethod.PATCH);
+        apiResponse = Request.execute(Endpoints.PERSON.get(), pathParams, person, ApiMethod.PATCH);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NO_CONTENT).log().body();
     }
 
     @Test(groups = "delete")
-    public void deleteAAccountTest() throws JsonProcessingException {
+    public void deleteAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), createdResponse.getId());
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, ApiMethod.DELETE);
+        apiResponse = Request.execute(Endpoints.PERSON.get(), pathParams, ApiMethod.DELETE);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NO_CONTENT).log().body();
     }
@@ -99,62 +102,63 @@ public class AccountTest extends CommonTest {
         pathParams.put(Endpoints.ID.get(), createdResponse.getId());
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, ApiMethod.DELETE);
+        apiResponse = Request.execute(Endpoints.PERSON.get(), pathParams, ApiMethod.DELETE);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NO_CONTENT).log().body();
     }
 
     @Test(groups = "get")
-    public void doNotGetAllAccountTest() throws JsonProcessingException {
+    public void doNotGetAllIndividualTest() throws JsonProcessingException {
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute("/Accoun", new HashMap<String, String>(), ApiMethod.GET);
+        apiResponse = Request.execute("/Inidivua", new HashMap<String, String>(), ApiMethod.GET);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NOT_FOUND).log().body();
     }
 
     @Test
-    public void doNotCreateAAccountTest() throws JsonProcessingException {
+    public void doNotCreateAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
-        Account account = new Account();
+        Person person = new Person();
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNTS.get(), pathParams, account, ApiMethod.POST);
+        apiResponse = Request.execute(Endpoints.PEOPLE.get(), pathParams, person, ApiMethod.POST);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_BAD_REQUEST).log().body();
     }
 
     @Test(groups = "get")
-    public void doNotGetAAccountTest() throws JsonProcessingException {
+    public void doNotGetAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), " ");
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, ApiMethod.GET);
+        apiResponse = Request.execute(Endpoints.PERSON.get(), pathParams, ApiMethod.GET);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NOT_FOUND).log().body();
     }
 
     @Test(groups = "update")
-    public void doNotUpdateAAccountTest() throws JsonProcessingException {
+    public void doNotUpdateAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), " ");
-        Account account = new Account();
-        account.setName("Updated Account");
+        Person person = new Person();
+        person.setFirstName("Pedrito");
+        person.setLastName("Fantasy 2");
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, account, ApiMethod.PATCH);
+        apiResponse = Request.execute(Endpoints.PERSON.get(), pathParams, person, ApiMethod.PATCH);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NOT_FOUND).log().body();
     }
 
     @Test(groups = {"delete", "badDelete"})
-    public void doNotDeleteAAccountTest() throws JsonProcessingException {
+    public void doNotDeleteAIndividualTest() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), " ");
         ApiResponse apiResponse;
 
-        apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, ApiMethod.DELETE);
+        apiResponse = Request.execute(Endpoints.PERSON.get(), pathParams, ApiMethod.DELETE);
 
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NOT_FOUND).log().body();
     }
