+++ src/test/java/rest/salesforce/com/tests/OrdersTest.java
@@ -5,10 +5,12 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package salesforce.tests;
+package rest.salesforce.com.tests;
 
 import api.*;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import api.ApiResponse;
+import salesforce.config.Request;
 import salesforce.config.Endpoints;
 import salesforce.entities.Account;
 import salesforce.entities.CreatedResponse;
@@ -37,7 +39,7 @@ public class OrdersTest extends CommonTest {
         Account account = new Account();
         account.setName("testAccount01");
         Map<String, String> pathParams = new HashMap<>();
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNTS.get(), pathParams, account, ApiMethod.POST);
+        ApiResponse apiResponse = Request.execute(Endpoints.ACCOUNTS.get(), pathParams, account, ApiMethod.POST);
         accountId = apiResponse.getBody(CreatedResponse.class).getId();
         Assert.assertEquals(apiResponse.getStatusCode(), HttpStatus.SC_CREATED);
     }
@@ -52,7 +54,7 @@ public class OrdersTest extends CommonTest {
         order.setEffectiveDate(date);
         order.setStatus("Draft");
         Map<String, String> pathParams = new HashMap<>();
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.ORDERS.get(), pathParams, order, ApiMethod.POST);
+        ApiResponse apiResponse = Request.execute(Endpoints.ORDERS.get(), pathParams, order, ApiMethod.POST);
         orderId = apiResponse.getBody(CreatedResponse.class).getId();
         Assert.assertEquals(apiResponse.getStatusCode(), HttpStatus.SC_CREATED);
     }
@@ -67,7 +69,7 @@ public class OrdersTest extends CommonTest {
         order.setEffectiveDate(date);
         order.setStatus("Draft");
         Map<String, String> pathParams = new HashMap<>();
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.ORDERS.get(), pathParams, order, ApiMethod.POST);
+        ApiResponse apiResponse = Request.execute(Endpoints.ORDERS.get(), pathParams, order, ApiMethod.POST);
         orderId = apiResponse.getBody(CreatedResponse.class).getId();
         Assert.assertEquals(apiResponse.getStatusCode(), HttpStatus.SC_CREATED);
     }
@@ -83,7 +85,7 @@ public class OrdersTest extends CommonTest {
         order.setStatus("Draft");
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), orderId);
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.ORDER.get(), pathParams, order, ApiMethod.PATCH);
+        ApiResponse apiResponse = Request.execute(Endpoints.ORDER.get(), pathParams, order, ApiMethod.PATCH);
         Assert.assertEquals(apiResponse.getStatusCode(), HttpStatus.SC_NO_CONTENT);
     }
 
@@ -91,7 +93,7 @@ public class OrdersTest extends CommonTest {
     public void deleteOrder() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), orderId);
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.ORDER.get(), pathParams, ApiMethod.DELETE);
+        ApiResponse apiResponse = Request.execute(Endpoints.ORDER.get(), pathParams, ApiMethod.DELETE);
         Assert.assertEquals(apiResponse.getStatusCode(), HttpStatus.SC_NO_CONTENT);
     }
 
@@ -99,7 +101,7 @@ public class OrdersTest extends CommonTest {
     public void deleteAccount() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), accountId);
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.ACCOUNT.get(), pathParams, ApiMethod.DELETE);
+        ApiResponse apiResponse = Request.execute(Endpoints.ACCOUNT.get(), pathParams, ApiMethod.DELETE);
         Assert.assertEquals(apiResponse.getStatusCode(), HttpStatus.SC_NO_CONTENT);
     }
 }
