+++ src/test/java/rest/salesforce/com/tests/ProductTest.java
@@ -5,10 +5,10 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package salesforce.tests;
+package rest.salesforce.com.tests;
 
 import api.ApiMethod;
-import api.ApiRequestManager;
+import salesforce.config.Request;
 import api.ApiResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import salesforce.config.Endpoints;
@@ -31,7 +31,7 @@ public class ProductTest extends CommonTest {
         product.setProductCode("FP1");
         product.setDescription("The description of the product");
         ApiResponse apiResponse;
-        apiResponse = ApiRequestManager.execute(Endpoints.PRODUCTS.get(), pathParams, product, ApiMethod.POST);
+        apiResponse = Request.execute(Endpoints.PRODUCTS.get(), pathParams, product, ApiMethod.POST);
         createdResponse = apiResponse.getResponse().as(CreatedResponse.class);
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_CREATED);
     }
@@ -41,13 +41,13 @@ public class ProductTest extends CommonTest {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), createdResponse.getId());
         ApiResponse apiResponse;
-        apiResponse = ApiRequestManager.execute(Endpoints.PRODUCT.get(), pathParams, ApiMethod.DELETE);
+        apiResponse = Request.execute(Endpoints.PRODUCT.get(), pathParams, ApiMethod.DELETE);
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NO_CONTENT);
     }
 
     @Test
     public void getProducts() throws JsonProcessingException {
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.PRODUCTS.get(), new HashMap<>(), ApiMethod.GET);
+        ApiResponse apiResponse = Request.execute(Endpoints.PRODUCTS.get(), new HashMap<>(), ApiMethod.GET);
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_OK);
     }
 
@@ -59,7 +59,7 @@ public class ProductTest extends CommonTest {
         product.setProductCode("FP1");
         product.setDescription("The description of the product");
         ApiResponse apiResponse;
-        apiResponse = ApiRequestManager.execute(Endpoints.PRODUCTS.get(), pathParams, product, ApiMethod.POST);
+        apiResponse = Request.execute(Endpoints.PRODUCTS.get(), pathParams, product, ApiMethod.POST);
         createdResponse = apiResponse.getResponse().as(CreatedResponse.class);
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_CREATED);
     }
@@ -68,7 +68,7 @@ public class ProductTest extends CommonTest {
     public void getAProduct() throws JsonProcessingException {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), createdResponse.getId());
-        ApiResponse apiResponse = ApiRequestManager.execute(Endpoints.PRODUCT.get(), pathParams, ApiMethod.GET);
+        ApiResponse apiResponse = Request.execute(Endpoints.PRODUCT.get(), pathParams, ApiMethod.GET);
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_OK);
     }
 
@@ -80,7 +80,7 @@ public class ProductTest extends CommonTest {
         product.setProductCode("FP1-1");
         product.setDescription("The description of the product with the description changed");
         ApiResponse apiResponse;
-        apiResponse = ApiRequestManager.execute(Endpoints.PRODUCT.get(), pathParams, product, ApiMethod.PATCH);
+        apiResponse = Request.execute(Endpoints.PRODUCT.get(), pathParams, product, ApiMethod.PATCH);
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NO_CONTENT);
     }
 
@@ -89,7 +89,7 @@ public class ProductTest extends CommonTest {
         Map<String, String> pathParams = new HashMap<>();
         pathParams.put(Endpoints.ID.get(), createdResponse.getId());
         ApiResponse apiResponse;
-        apiResponse = ApiRequestManager.execute(Endpoints.PRODUCT.get(), pathParams, ApiMethod.DELETE);
+        apiResponse = Request.execute(Endpoints.PRODUCT.get(), pathParams, ApiMethod.DELETE);
         apiResponse.getResponse().then().assertThat().statusCode(HttpStatus.SC_NO_CONTENT);
     }
 }
