+++ src/test/resources/features/campaign.feature
@@ -2,13 +2,12 @@ Feature: Campaign
 
   @GetCampaigns
   Scenario: Get all Campaigns
-  Given
     When I set the "/Campaign" endpoint and send the request
     Then the response status code should be "200"
 
   @CreateCampaign
   Scenario Outline: Create a Campaign
-    Given I set the pathParams and body to request
+    Given I set the post request
       | name     | <nameCampaign>     |
       | isActive | <isActiveCampaign> |
       | type     | <typeCampaign>     |
@@ -26,13 +25,13 @@ Feature: Campaign
 
   @GetCampaign
   Scenario: Get a Campaign
-    Given I set the pathParams to request
+    Given I set the "get" request
     When I set the "/Campaign/{id}" endpoint and send the request
     Then the response status code should be "200" with the "campaign" schema
 
   @UpdateCampaign
   Scenario Outline: Update a Product
-    Given I set the pathParams and updated body to request
+    Given I set the update request
       | name     | <nameCampaign>     |
       | isActive | <isActiveCampaign> |
       | type     | <typeCampaign>     |
@@ -50,6 +49,6 @@ Feature: Campaign
 
   @DeleteCampaign
   Scenario: Delete a Campaign
-    Given I set the pathParams to request
+    Given I set the "delete" request
     When I set the "/Campaign/{id}" endpoint and send the delete request
     Then the response status code should be "204"
