+++ src/test/resources/features/contact.feature
@@ -2,7 +2,7 @@ Feature: Contact
 
   @CreateContact
   Scenario Outline: Create contact with firstname and lastname values
-    Given I set the pathParams and body to request
+    Given I set the post request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
     When I set the "/Contact" endpoint and send the request with body
@@ -25,7 +25,7 @@ Feature: Contact
 
   @CreateContact
   Scenario Outline: Create contact with firstname, lastname and birthdate values
-    Given I set the pathParams and body to request
+    Given I set the post request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
       | birthdate | <birthdate> |
@@ -46,7 +46,7 @@ Feature: Contact
 
   @CreateContact
   Scenario Outline: Create contact with firstname, lastname and department values
-    Given I set the pathParams and body to request
+    Given I set the post request
       | firstname  | <firstname>  |
       | lastname   | <lastname>   |
       | department | <department> |
@@ -65,7 +65,7 @@ Feature: Contact
 
   @CreateContact
   Scenario Outline: Create contact with firstname, lastname and phone values
-    Given I set the pathParams and body to request
+    Given I set the post request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
       | phone     | <phone>     |
@@ -84,7 +84,7 @@ Feature: Contact
 
   @CreateContact
   Scenario Outline: Create contact with firstname, lastname and description values
-    Given I set the pathParams and body to request
+    Given I set the post request
       | firstname   | <firstname>   |
       | lastname    | <lastname>    |
       | description | <description> |
@@ -101,7 +101,7 @@ Feature: Contact
 
   @CreateContact
   Scenario Outline: Create contact with firstname, lastname and email values
-    Given I set the pathParams and body to request
+    Given I set the post request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
       | email     | <email>     |
@@ -119,13 +119,13 @@ Feature: Contact
 
   @DeleteContact
   Scenario: Delete a Contact
-    Given I set the pathParams to request
+    Given I set the "delete" request
     When I set the "/Contact/{id}" endpoint and send the delete request
     Then the response status code should be "204"
 
   @UpdateContact
   Scenario Outline: Update a Contact firstname and lastname
-    Given I set the pathParams and updated body to request
+    Given I set the update request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
     When I set the "/Contact/{id}" endpoint and send the request with updated body
@@ -145,7 +145,7 @@ Feature: Contact
 
   @UpdateContact
   Scenario Outline: Update a Contact birthday value
-    Given I set the pathParams and updated body to request
+    Given I set the update request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
       | birthdate | <birthdate> |
@@ -165,7 +165,7 @@ Feature: Contact
 
   @UpdateContact
   Scenario Outline: Update a Contact department value
-    Given I set the pathParams and updated body to request
+    Given I set the update request
       | firstname  | <firstname>  |
       | lastname   | <lastname>   |
       | department | <department> |
@@ -184,7 +184,7 @@ Feature: Contact
 
   @UpdateContact
   Scenario Outline: Update a Contact phone value
-    Given I set the pathParams and updated body to request
+    Given I set the update request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
       | phone     | <phone>     |
@@ -202,7 +202,7 @@ Feature: Contact
 
   @UpdateContact
   Scenario Outline: Update a Contact description value
-    Given I set the pathParams and updated body to request
+    Given I set the update request
       | firstname   | <firstname>   |
       | lastname    | <lastname>    |
       | description | <description> |
@@ -219,7 +219,7 @@ Feature: Contact
 
   @UpdateContact
   Scenario Outline: Update a Contact email value
-    Given I set the pathParams and updated body to request
+    Given I set the update request
       | firstname | <firstname> |
       | lastname  | <lastname>  |
       | email     | <email>     |
